Avoid signed-to-unsigned comparison

+++ num/__private/literals.h
@@ -15,7 +15,6 @@
 #pragma once
 
 #if _MSC_VER
-
 // Workaround for MSVC bug that doesn't constant-evaluate user-defined literals
 // in all cases:
 // https://developercommunity.visualstudio.com/t/User-defined-literals-not-constant-expre/10108165
@@ -26,15 +25,15 @@
 // that triggers a different MSVC bug when used with any unary/binary operator
 // in a templated function:
 // https://developercommunity.visualstudio.com/t/MSVC-Compiler-bug-with:-numeric-literal/10108160
-#define _sus__integer_literal(Name, T, PrimitiveT)              \
-  T inline constexpr operator""_##Name(unsigned long long val) { \
-    ::sus::check(val <= T::MAX_PRIMITIVE);                       \
-    return T(static_cast<decltype(T::primitive_value)>(val));    \
+#define _sus__integer_literal(Name, T, PrimitiveT)                          \
+  T inline constexpr operator""_##Name(unsigned long long val) {            \
+    ::sus::check(val <= static_cast<unsigned long long>(T::MAX_PRIMITIVE)); \
+    return T(static_cast<decltype(T::primitive_value)>(val));               \
   }
 #else
-#define _sus__integer_literal(Name, T, PrimitiveT)              \
+#define _sus__integer_literal(Name, T, PrimitiveT)               \
   T inline consteval operator""_##Name(unsigned long long val) { \
-    if (val > T::MAX_PRIMITIVE)                                  \
+    if (val > static_cast<unsigned long long>(T::MAX_PRIMITIVE)) \
       throw "Integer literal out of bounds for ##T##";           \
     return T(static_cast<decltype(T::primitive_value)>(val));    \
   }
